chore: Refactoring ColumnPinning into folder with separate base, types, and utils
