+++ packages/table-core/src/core/table.ts
@@ -27,7 +27,7 @@ import { ColumnFaceting } from '../features/column-faceting/ColumnFaceting'
 import { ColumnFiltering } from '../features/column-filtering/ColumnFiltering'
 import { ColumnGrouping } from '../features/column-grouping/ColumnGrouping'
 import { ColumnOrdering } from '../features/column-ordering/ColumnOrdering'
-import { ColumnPinning } from '../features/ColumnPinning'
+import { ColumnPinning } from '../features/column-pinning/ColumnPinning'
 import { ColumnSizing } from '../features/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
 import { GlobalFaceting } from '../features/GlobalFaceting'
